+++ rust/dbn-cli/src/lib.rs
@@ -10,10 +10,11 @@ use clap::{ArgAction, Parser, ValueEnum};
 
 use dbn::{
     enums::{Compression, Encoding},
-    VersionUpgradePolicy,
+    Schema, VersionUpgradePolicy,
 };
 
 pub mod encode;
+pub mod filter;
 
 /// How the output of the `dbn` command will be encoded.
 #[derive(Clone, Copy, Debug, ValueEnum)]
@@ -118,7 +119,6 @@ pub struct Args {
         short = 'l',
         long = "limit",
         value_name = "NUM_RECORDS",
-        conflicts_with = "should_output_metadata",
         help = "Limit the number of records in the output to the specified number"
     )]
     pub limit: Option<NonZeroU64>,
@@ -149,6 +149,12 @@ pub struct Args {
         requires = "input_fragment"
     )]
     pub input_dbn_version_override: Option<u8>,
+    #[clap(
+        long = "schema",
+        help = "Only encode records of this schema. This is particularly useful for transcoding mixed-schema DBN to CSV, which doesn't support mixing schemas",
+        value_name = "SCHEMA"
+    )]
+    pub schema_filter: Option<Schema>,
 }
 
 impl Args {
@@ -173,6 +179,10 @@ impl Args {
         } else {
             VersionUpgradePolicy::AsIs
         }
+    }
+
+    pub fn input_version(&self) -> u8 {
+        self.input_dbn_version_override.unwrap_or(dbn::DBN_VERSION)
     }
 }
 
@@ -228,8 +238,7 @@ pub fn output_from_args(args: &Args) -> anyhow::Result<Box<dyn io::Write>> {
 
 fn open_output_file(path: &PathBuf, force: bool) -> anyhow::Result<File> {
     let mut options = File::options();
-    options.write(true);
-    options.truncate(true);
+    options.write(true).truncate(true);
     if force {
         options.create(true);
     } else if path.exists() {